feat: Complete the files endpoints and service layer

+++ src/main/java/com/example/superduperdrive/model/Note.java
@@ -2,23 +2,23 @@ package com.example.superduperdrive.model;
 
 public class Note {
 
-    private int noteId;
+    private Long noteId;
     private String noteTitle;
     private String noteDescription;
 
     public Note() { }
 
-    public Note(int noteId, String noteTitle, String noteDescription) {
+    public Note(Long noteId, String noteTitle, String noteDescription) {
         this.noteId = noteId;
         this.noteTitle = noteTitle;
         this.noteDescription = noteDescription;
     }
 
-    public int getNoteId() {
+    public Long getNoteId() {
         return noteId;
     }
 
-    public void setNoteId(int noteId) {
+    public void setNoteId(Long noteId) {
         this.noteId = noteId;
     }
